requirements.txt
astroid==2.5.2
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
DateTime==4.3
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pycodestyle==2.7.0
pydocstyle==6.0.0
pyflakes==2.3.1
pylama==7.7.1
pylint==2.7.4
python-dotenv==0.16.0
pytz==2021.1
requests==2.25.1
six==1.15.0
snowballstemmer==2.1.0
toml==0.10.2
tz==0.2.2
urllib3==1.26.4
websocket-client==0.56.0
wrapt==1.12.1
zope.interface==5.3.0
